Move keyboard layouts to its own file.
Moved keyboard layout definition to its own file (keyboard_layouts.h).
